docs/modul-3/woche-8/code/rezeptbuch/backend/requirements.txt
fastapi~=0.133.0
uvicorn~=0.41.0
psycopg2-binary~=2.9.0
motor~=3.6.0
